services/llm-gateway-py/requirements.txt
# Core dependencies
fastapi==0.104.1
uvicorn==0.24.0
grpcio>=1.65.0
grpcio-tools>=1.65.0
grpcio-reflection>=1.65.0
protobuf>=4.25.1

# LangChain and AI/ML
langchain>=0.2.0
langchain-openai>=0.1.8
langchain-community>=0.2.0
langchain-pinecone>=0.2.6
openai>=1.35.0
tavily-python>=0.3.8

# Vector database and embeddings
pinecone-client==4.1.2
chromadb==0.5.5
sentence-transformers==3.0.1

# Data processing
pydantic==2.8.2
python-multipart==0.0.9
python-dotenv==1.0.1

# Monitoring and logging
structlog==23.2.0
prometheus-client==0.19.0

# Security and validation
bcrypt==4.1.2
passlib==1.7.4
PyJWT==2.8.0

# Development and testing
pytest==7.4.3
pytest-asyncio==0.21.1
black==23.11.0
isort==5.12.0
mypy==1.7.1
